third_party/WebKit/LayoutTests/media/video-controls-muted-video-can-unmute.html

<!DOCTYPE html>
<title>Test that muted video has an unmute button.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<video controls></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    video.muted = true;

    video.oncanplaythrough = t.step_func_done(function() {
        assert_true(video.muted);
        var muteButton = mediaControlsButton(video, "mute-button");
        // Make sure that the mute button is visible.
        assert_not_equals(getComputedStyle(muteButton).display, "none");
    });

    video.src = "content/test.ogv";
});
</script>
